+++ app/keepers/keepers.go
@@ -16,7 +16,6 @@ import (
 	"github.com/CosmWasm/wasmd/x/wasm"
 	wasmkeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
 	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
-	"github.com/babylonlabs-io/babylon/testutil/addr"
 	mintkeeper "github.com/babylonlabs-io/babylon/x/mint/keeper"
 	minttypes "github.com/babylonlabs-io/babylon/x/mint/types"
 	"github.com/cosmos/cosmos-sdk/baseapp"
@@ -219,7 +218,7 @@ func (ak *AppKeepers) InitKeepers(
 		maccPerms,
 		authcodec.NewBech32Codec(appparams.Bech32PrefixAccAddr),
 		appparams.Bech32PrefixAccAddr,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	bankKeeper := bankkeeper.NewBaseKeeper(
@@ -227,7 +226,7 @@ func (ak *AppKeepers) InitKeepers(
 		runtime.NewKVStoreService(keys[banktypes.StoreKey]),
 		accountKeeper,
 		blockedAddress,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 		logger,
 	)
 
@@ -236,7 +235,7 @@ func (ak *AppKeepers) InitKeepers(
 		runtime.NewKVStoreService(keys[stakingtypes.StoreKey]),
 		accountKeeper,
 		bankKeeper,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 		authcodec.NewBech32Codec(appparams.Bech32PrefixValAddr),
 		authcodec.NewBech32Codec(appparams.Bech32PrefixConsAddr),
 	)
@@ -247,7 +246,7 @@ func (ak *AppKeepers) InitKeepers(
 		runtime.NewKVStoreService(keys[epochingtypes.StoreKey]),
 		bankKeeper,
 		stakingKeeper,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	checkpointingKeeper := checkpointingkeeper.NewKeeper(
@@ -272,7 +271,7 @@ func (ak *AppKeepers) InitKeepers(
 	ak.ConsensusParamsKeeper = consensusparamkeeper.NewKeeper(
 		appCodec,
 		runtime.NewKVStoreService(keys[consensusparamtypes.StoreKey]),
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 		runtime.EventService{},
 	)
 	bApp.SetParamStore(ak.ConsensusParamsKeeper.ParamsStore)
@@ -302,7 +301,7 @@ func (ak *AppKeepers) InitKeepers(
 	ak.CircuitKeeper = circuitkeeper.NewKeeper(
 		appCodec,
 		runtime.NewKVStoreService(keys[circuittypes.StoreKey]),
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 		ak.AccountKeeper.AddressCodec(),
 	)
 	bApp.SetCircuitBreaker(&ak.CircuitKeeper)
@@ -314,7 +313,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.AccountKeeper,
 		ak.BankKeeper,
 		authtypes.FeeCollectorName,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	ak.DistrKeeper = distrkeeper.NewKeeper(
@@ -324,7 +323,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.BankKeeper,
 		ak.StakingKeeper,
 		authtypes.FeeCollectorName,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	// set up incentive keeper
@@ -334,7 +333,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.BankKeeper,
 		ak.AccountKeeper,
 		&epochingKeeper,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 		authtypes.FeeCollectorName,
 	)
 
@@ -343,7 +342,7 @@ func (ak *AppKeepers) InitKeepers(
 		encodingConfig.Amino,
 		runtime.NewKVStoreService(keys[slashingtypes.StoreKey]),
 		ak.StakingKeeper,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	ak.CrisisKeeper = crisiskeeper.NewKeeper(
@@ -352,7 +351,7 @@ func (ak *AppKeepers) InitKeepers(
 		invCheckPeriod,
 		ak.BankKeeper,
 		authtypes.FeeCollectorName,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 		ak.AccountKeeper.AddressCodec(),
 	)
 
@@ -374,7 +373,7 @@ func (ak *AppKeepers) InitKeepers(
 		appCodec,
 		homePath,
 		bApp,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	ak.AuthzKeeper = authzkeeper.NewKeeper(
@@ -395,7 +394,7 @@ func (ak *AppKeepers) InitKeepers(
 		// `MsgIBCSoftwareUpgrade` and `MsgRecoverClient`
 		// https://github.com/cosmos/ibc-go/releases/tag/v8.0.0
 		// Gov is the proper authority for those types of messages
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	// register the proposal types
@@ -422,7 +421,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.DistrKeeper,
 		bApp.MsgServiceRouter(),
 		govConfig,
-		addr.AccGov.String())
+		appparams.AccGov.String())
 
 	ak.GovKeeper = *govKeeper.SetHooks(
 		govtypes.NewMultiGovHooks(
@@ -435,7 +434,7 @@ func (ak *AppKeepers) InitKeepers(
 		runtime.NewKVStoreService(keys[btclightclienttypes.StoreKey]),
 		*btcConfig,
 		&ak.IncentiveKeeper,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	btcCheckpointKeeper := btccheckpointkeeper.NewKeeper(
@@ -446,7 +445,7 @@ func (ak *AppKeepers) InitKeepers(
 		&checkpointingKeeper,
 		&ak.IncentiveKeeper,
 		&powLimit,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	ak.IBCFeeKeeper = ibcfeekeeper.NewKeeper(
@@ -467,7 +466,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.AccountKeeper,
 		ak.BankKeeper,
 		scopedTransferKeeper,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	ak.MonitorKeeper = monitorkeeper.NewKeeper(
@@ -500,7 +499,7 @@ func (ak *AppKeepers) InitKeepers(
 		&btcCheckpointKeeper,
 		&ak.IncentiveKeeper,
 		btcNetParams,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	// set up finality keeper
@@ -510,7 +509,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.BTCStakingKeeper,
 		ak.IncentiveKeeper,
 		ak.CheckpointingKeeper,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 	)
 
 	// create evidence keeper with router
@@ -544,7 +543,7 @@ func (ak *AppKeepers) InitKeepers(
 		homePath,
 		wasmConfig,
 		WasmCapabilities(),
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 		wasmOpts...,
 	)
 
@@ -559,7 +558,7 @@ func (ak *AppKeepers) InitKeepers(
 		appCodec,
 		runtime.NewKVStoreService(keys[ibcwasmtypes.StoreKey]),
 		ak.IBCKeeper.ClientKeeper,
-		addr.AccGov.String(),
+		appparams.AccGov.String(),
 		ibcWasmConfig,
 		bApp.GRPCQueryRouter(),
 	)