+++ CHANGELOG.md
@@ -46,6 +46,8 @@ virtual block periods for each finality provider that has delegations and reward
 - [#316](https://github.com/babylonlabs-io/babylon/pull/316) Add testnet upgrade data
 - [#326](https://github.com/babylonlabs-io/babylon/pull/326) docs: btcstaking:
 Update btcstaking module docs to include EOI
+- [#348](https://github.com/babylonlabs-io/babylon/pull/348) refactory `PrivateSigner`
+and module account vars in appparams
 
 ### Bug fixes